docs(Table): improve MyTable & dedupe usage
