+++ apps/docs/src/app/examples/tables/CustomCellsTable.tsx
@@ -3,22 +3,14 @@ import {
   HvCellProps,
   HvIconButton,
   HvStatusIcon,
-  HvTable,
-  HvTableBody,
-  HvTableCell,
   HvTableColumnConfig,
-  HvTableContainer,
-  HvTableHead,
-  HvTableHeader,
-  HvTableRow,
-  HvTableSection,
   HvTag,
   HvTypography,
-  useHvTable,
 } from "@hitachivantara/uikit-react-core";
 import { Delete } from "@hitachivantara/uikit-react-icons";
 
-import { getPriorityColor, makeData, type AssetEvent } from "./makeData";
+import { makeData, type AssetEvent } from "./makeData";
+import { MyTable } from "./MyTable";
 
 export default function Demo() {
   const [data] = useState(() => makeData(8));
@@ -87,62 +79,8 @@ export default function Demo() {
     [],
   );
 
-  return <MyTable data={data} columns={columns} />;
+  return <MyTable<AssetEvent> data={data} columns={columns} />;
 }
-
-/** A simple generic client-side table. */
-export const MyTable = <T extends object>(props: {
-  columns: HvTableColumnConfig<T>[];
-  data: T[] | undefined;
-}) => {
-  const { columns, data } = props;
-
-  const table = useHvTable<T>({ columns, data });
-
-  return (
-    <HvTableSection>
-      <HvTableContainer className="max-h-500px">
-        <HvTable {...table.getTableProps()}>
-          <HvTableHead {...table.getTableHeadProps?.()}>
-            {table.headerGroups.map((headerGroup) => (
-              <HvTableRow
-                {...headerGroup.getHeaderGroupProps()}
-                key={headerGroup.getHeaderGroupProps().key}
-              >
-                {headerGroup.headers.map((col) => (
-                  <HvTableHeader
-                    {...col.getHeaderProps()}
-                    key={col.getHeaderProps().key}
-                  >
-                    {col.render("Header")}
-                  </HvTableHeader>
-                ))}
-              </HvTableRow>
-            ))}
-          </HvTableHead>
-          <HvTableBody {...table.getTableBodyProps()}>
-            {table.rows.map((row) => {
-              table.prepareRow(row);
-              return (
-                <HvTableRow {...row.getRowProps()} key={row.getRowProps().key}>
-                  {row.cells.map((cell) => (
-                    <HvTableCell
-                      className="text-nowrap"
-                      {...cell.getCellProps()}
-                      key={cell.getCellProps().key}
-                    >
-                      {cell.render("Cell")}
-                    </HvTableCell>
-                  ))}
-                </HvTableRow>
-              );
-            })}
-          </HvTableBody>
-        </HvTable>
-      </HvTableContainer>
-    </HvTableSection>
-  );
-};
 
 const TitleDescription = ({ row }: HvCellProps<AssetEvent>) => (
   <div className="p-xxs">
@@ -156,17 +94,26 @@ const TitleDescription = ({ row }: HvCellProps<AssetEvent>) => (
 const ProgressBar = ({ row }: HvCellProps<AssetEvent>) => (
   <div className="flex flex-col">
     <HvTypography variant="caption2">
-      {row.original.riskScore * 10}%
+      {`${row.original.riskScore * 10}%`}
     </HvTypography>
-    <div className="flex h-4px">
+    <div className="flex h-4px bg-border">
       <div
         className="bg-text"
         style={{ width: `${row.original.riskScore * 10}%` }}
       />
-      <div
-        className="bg-border"
-        style={{ width: `${100 - row.original.riskScore * 10}%` }}
-      />
     </div>
   </div>
 );
+
+function getPriorityColor(priority: AssetEvent["priority"]) {
+  switch (priority) {
+    case "High":
+      return "negative";
+    case "Medium":
+      return "warning";
+    case "Low":
+      return "positive";
+    default:
+      return "neutral";
+  }
+}
