+++ apps/docs/src/app/examples/tables/SimpleTable.tsx
@@ -1,20 +1,12 @@
 import { useMemo, useState } from "react";
 import {
   HvIconButton,
-  HvTable,
-  HvTableBody,
-  HvTableCell,
   HvTableColumnConfig,
-  HvTableContainer,
-  HvTableHead,
-  HvTableHeader,
-  HvTableRow,
-  HvTableSection,
-  useHvTable,
 } from "@hitachivantara/uikit-react-core";
 import { Delete } from "@hitachivantara/uikit-react-icons";
 
 import { makeData, type AssetEvent } from "./makeData";
+import { MyTable } from "./MyTable";
 
 export default function Demo() {
   const [data] = useState(() => makeData(8));
@@ -45,59 +37,5 @@ export default function Demo() {
     [],
   );
 
-  return <MyTable data={data} columns={columns} />;
+  return <MyTable<AssetEvent> data={data} columns={columns} />;
 }
-
-/** A simple generic client-side table. */
-export const MyTable = <T extends object>(props: {
-  columns: HvTableColumnConfig<T>[];
-  data: T[] | undefined;
-}) => {
-  const { columns, data } = props;
-
-  const table = useHvTable<T>({ columns, data });
-
-  return (
-    <HvTableSection>
-      <HvTableContainer className="max-h-500px">
-        <HvTable {...table.getTableProps()}>
-          <HvTableHead {...table.getTableHeadProps?.()}>
-            {table.headerGroups.map((headerGroup) => (
-              <HvTableRow
-                {...headerGroup.getHeaderGroupProps()}
-                key={headerGroup.getHeaderGroupProps().key}
-              >
-                {headerGroup.headers.map((col) => (
-                  <HvTableHeader
-                    {...col.getHeaderProps()}
-                    key={col.getHeaderProps().key}
-                  >
-                    {col.render("Header")}
-                  </HvTableHeader>
-                ))}
-              </HvTableRow>
-            ))}
-          </HvTableHead>
-          <HvTableBody {...table.getTableBodyProps()}>
-            {table.rows.map((row) => {
-              table.prepareRow(row);
-              return (
-                <HvTableRow {...row.getRowProps()} key={row.getRowProps().key}>
-                  {row.cells.map((cell) => (
-                    <HvTableCell
-                      className="text-nowrap"
-                      {...cell.getCellProps()}
-                      key={cell.getCellProps().key}
-                    >
-                      {cell.render("Cell")}
-                    </HvTableCell>
-                  ))}
-                </HvTableRow>
-              );
-            })}
-          </HvTableBody>
-        </HvTable>
-      </HvTableContainer>
-    </HvTableSection>
-  );
-};
