+++ apps/docs/src/app/examples/tables/makeData.ts
@@ -4,7 +4,7 @@ const makeEvent = (i: number) => ({
   id: `${i + 1}`,
   name: `Event ${i + 1}`,
   description: `Description for event ${i + 1}`,
-  createdDate: new Date("2020-03-20").toISOString().slice(0, 10),
+  createdDate: new Date(2020, 3, 10 + i).toISOString().slice(0, 10),
   eventType: "Anomaly detection",
   status: getOption(["Closed", "Open"] as const, i),
   riskScore: (i % 100) + 1,
@@ -15,16 +15,3 @@ const makeEvent = (i: number) => ({
 export type AssetEvent = ReturnType<typeof makeEvent>;
 
 export const makeData = (len = 10) => [...Array(len).keys()].map(makeEvent);
-
-export const getPriorityColor = (priority: string) => {
-  switch (priority.toLowerCase()) {
-    case "high":
-      return "negative";
-    case "medium":
-      return "warning";
-    case "low":
-      return "positive";
-    default:
-      return "neutral";
-  }
-};
